WIP conversion to use powermenu from rofi modi

+++ .bin/rofi_powermenu.sh
@@ -1,21 +1,25 @@
 #!/bin/bash
 
 menu="lock
-logout suspend
+logout
+suspend
 hibernate
 reboot
 shutdown
 "
 
-args="--quality 100"
-opt=$(rofi -dmenu <<< $menu | cut -d' ' -f1)
-
-case "$opt" in
-    lock) i3scrlock ;;
-    logout) gnome-session-quit --logout ;;
-    reboot) systemctl reboot ;;
-    shutdown) systemctl poweroff ;;
-    suspend) i3locksusp ;;
-    *) exit 1 ;;
-esac
-
+if [ -z "$@" ]; then
+    cat <<< $menu  
+else
+    option=$@
+    if [ ! -z "$option" ]; then
+        case "$option" in
+            lock) i3scrlock ;;
+            logout) gnome-session-quit --logout ;;
+            reboot) systemctl reboot ;;
+            shutdown) systemctl poweroff ;;
+            suspend) i3locksusp ;;
+            *) exit 1 ;;
+        esac
+    fi
+fi
